refactor out common boilerplate code
