+++ src/core/ir/jit/mod.rs
@@ -6,3 +6,6 @@ pub use builder::*;
 pub use model::*;
 pub use store::*;
 pub use synth::*;
+
+// NOTE: Only used in tests and benchmarks
+pub mod common;
